+++ src/components/Buttons/LogoutButton.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { ListItem, ListItemIcon, ListItemText, IconButton } from '@mui/material';
 import { Logout } from '@mui/icons-material';
-import { AuthContext } from '../context/AuthContext';
+import { AuthContext } from '../../context/AuthContext';
 
 const LogoutButton: React.FC = () => {
     const { logout } = useContext(AuthContext) || { logout: () => {} };
